add metal capture unused commands test
This create 2 textures, 2 renderPassStates, 2 command buffers
but only submits one. We then look at the trace. We see the
trace only shows 1 texture, 1 renderPassStates, 1 command buffer.

In other words, this suggests Metal is not recording API calls.
It's recording the command buffer itself and that command buffer's
referenced resources.

If it was recording API calls then we'd see the 2nd texture,
we'd see the 2nd renderPassStates, and we'd see the commands
added to the 2nd command buffer.